fix crf dtype bug

+++ delta/utils/loss/loss_utils_test.py
@@ -218,8 +218,8 @@ class LossUtilTest(tf.test.TestCase):
                             [0.3, 0.2, 0.5], [0.6, 0.2, 0.2]]],
                           dtype=np.float32)  # [1,5,3]
       trans_params = tf.fill([3, 3], 0.5, name='trans_params')
-      labels = np.asarray([[0, 1, 2, 0, 1]])  # shape=[1,5]
-      sequence_lengths = np.asarray([5])  # shape=[1,]
+      labels = np.asarray([[0, 1, 2, 0, 1]], dtype=np.int32)  # shape=[1,5]
+      sequence_lengths = np.asarray([5], dtype=np.int32)  # shape=[1,]
       loss, _ = loss_utils.crf_log_likelihood(
           tf.constant(logits), tf.constant(labels),
           tf.constant(sequence_lengths), trans_params)
